requirements.txt
tweepy==4.14.0
spacy==3.7.2
nltk==3.8.1
python-dateutil==2.8.2
SQLAlchemy==2.0.23
schedule==1.2.1
Flask==3.0.0
python-telegram-bot==20.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz 